Add files via upload
Usually, text data consumes more time in cleaning. The nature of text data is fragile which makes it prone to errors. Working with text data is tiring and also different problems will require different analysis to be performed. Not all texts need to be cleaned in the same manner. Some need extra cleaning and some don't.